# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Suite Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  tokenBalance: number;
  /** @nullable */
  planId: string | null;
  /** @nullable */
  planName?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Tool {
  id: string;
  label: string;
  category: string;
  /** @nullable */
  description?: string | null;
  tokenCost: number;
  /** @nullable */
  isNew?: boolean | null;
  /** @nullable */
  isPro?: boolean | null;
  /** @nullable */
  usageCount?: number | null;
  /** @nullable */
  n8nWebhookUrl?: string | null;
}

export interface ToolCategory {
  id: string;
  label: string;
  toolCount: number;
}

/**
 * @nullable
 */
export type GenerateInputExtraInputs = { [key: string]: unknown } | null;

export interface GenerateInput {
  toolId: string;
  prompt: string;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  extraInputs?: GenerateInputExtraInputs;
}

export interface GenerateResponse {
  text: string;
  tokensUsed: number;
  model?: string;
}

export interface AiModel {
  id: string;
  name: string;
  provider: string;
  /** @nullable */
  description?: string | null;
}

export type UsageStatsTopToolsItem = {
  toolId: string;
  label: string;
  count: number;
};

export interface UsageStats {
  totalGenerations: number;
  tokensUsed: number;
  tokenBalance: number;
  topTools: UsageStatsTopToolsItem[];
}

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  role: string;
  tokenBalance: number;
  /** @nullable */
  planId: string | null;
  /** @nullable */
  planName?: string | null;
  createdAt: string;
  totalGenerations?: number;
  favoriteToolIds?: string[];
}

export interface UserProfileUpdate {
  name?: string;
  email?: string;
}

export interface GenerationRecord {
  id: number;
  toolId: string;
  toolLabel: string;
  /** @nullable */
  prompt?: string | null;
  /** @nullable */
  output?: string | null;
  tokensUsed: number;
  /** @nullable */
  model?: string | null;
  createdAt: string;
}

export type PlanInterval = typeof PlanInterval[keyof typeof PlanInterval];


export const PlanInterval = {
  monthly: 'monthly',
  yearly: 'yearly',
  lifetime: 'lifetime',
} as const;

export interface Plan {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  interval: PlanInterval;
  tokenAllowance: number;
  features: string[];
  /** @nullable */
  isPopular?: boolean | null;
  /** @nullable */
  woocommerceProductId?: number | null;
}

export interface SubscribeInput {
  planId: string;
}

export interface SubscribeResponse {
  success: boolean;
  checkoutUrl: string;
  /** @nullable */
  message?: string | null;
}

export type DashboardStatsGenerationsByDayItem = {
  date: string;
  count: number;
};

export interface DashboardStats {
  tokenBalance: number;
  tokensUsed: number;
  totalGenerations: number;
  planName: string;
  favoriteCount: number;
  generationsByDay?: DashboardStatsGenerationsByDayItem[];
}

export type AdminStatsUsersByPlanItem = {
  planName: string;
  count: number;
};

export interface AdminStats {
  totalUsers: number;
  totalGenerations: number;
  totalTokensUsed: number;
  activeToday: number;
  revenueTotal: number;
  usersByPlan?: AdminStatsUsersByPlanItem[];
}

export interface AdminUserUpdate {
  tokenBalance?: number;
  planId?: string;
  role?: string;
  isActive?: boolean;
}

export type ListToolsParams = {
category?: string;
search?: string;
};

export type GetUserHistoryParams = {
limit?: number;
offset?: number;
};

export type ListUsersParams = {
limit?: number;
offset?: number;
search?: string;
};