# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Suite Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Tools ──────────────────────────────────────────────────────────────────
  /tools:
    get:
      operationId: listTools
      tags: [tools]
      summary: List all AI tools
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tool"
  /tools/{toolId}:
    get:
      operationId: getTool
      tags: [tools]
      summary: Get a single tool by ID
      parameters:
        - name: toolId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tool details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ai/generate:
    post:
      operationId: generateAi
      tags: [ai]
      summary: Generate AI output (non-streaming)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInput"
      responses:
        "200":
          description: Generated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"
        "402":
          description: Insufficient tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /plans/subscribe:
    post:
      operationId: subscribeToPlan
      tags: [plans]
      summary: Subscribe to a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeInput"
      responses:
        "200":
          description: Subscription result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscribeResponse"

  # ─── Dashboard summary ───────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{userId}:
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update a user (tokens, plan, status)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ─── Auth ─────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, tokenBalance, planId, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        tokenBalance:
          type: integer
        planId:
          type: ["string", "null"]
        planName:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── Tools ────────────────────────────────────────────────────────────────
    Tool:
      type: object
      required: [id, label, category, tokenCost]
      properties:
        id:
          type: string
        label:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        tokenCost:
          type: integer
        isNew:
          type: ["boolean", "null"]
        isPro:
          type: ["boolean", "null"]
        usageCount:
          type: ["integer", "null"]
        n8nWebhookUrl:
          type: ["string", "null"]
...
    GenerateInput:
      type: object
      required: [toolId, prompt]
      properties:
        toolId:
          type: string
        prompt:
          type: string
        model:
          type: ["string", "null"]
        extraInputs:
          type: ["object", "null"]
          additionalProperties: true
    GenerateResponse:
      type: object
      required: [text, tokensUsed]
      properties:
        text:
          type: string
        tokensUsed:
          type: integer
        model:
          type: string
...
    SubscribeInput:
      type: object
      required: [planId]
      properties:
        planId:
          type: string
    SubscribeResponse:
      type: object
      required: [success, checkoutUrl]
      properties:
        success:
          type: boolean
        checkoutUrl:
          type: string
        message:
          type: ["string", "null"]

    # ─── Dashboard ────────────────────────────────────────────────────────────
...
    AdminUserUpdate:
      type: object
      properties:
        tokenBalance:
          type: integer
        planId:
          type: string
        role:
          type: string
        isActive:
          type: boolean