# lib/api-zod/src/generated/types/generateInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Suite Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateInputExtraInputs } from './generateInputExtraInputs';

export interface GenerateInput {
  toolId: string;
  prompt: string;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  extraInputs?: GenerateInputExtraInputs;
}