# lib/api-zod/src/generated/types/generateResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Suite Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateResponse {
  text: string;
  tokensUsed: number;
  model?: string;
}